requirements.txt
attrs
tgalice>=0.2.11
dnspython
pymongo
pyyaml
regex
